refactor(stats): extract useStatsData hook and split UI components
Extracts heavy state logic and calculations into `useStatsData` hook.
Splits the main UI rendering in `StatsPage` into smaller sub-components (`SummaryGrid`, `PrefectureSection`, `VisitCalendar`).
This improves the readability and maintainability of `StatsPage`.
