+++ logs/code_review_result.md
@@ -1,0 +1,18 @@
+# Code Review Result
+
+**Analysis and Reasoning:**
+
+1.  **User's Goal:** The user wants to refactor the `StatsPage` component (which has an overly long default export) by extracting calculation logic and UI sections into separate hooks and components to improve maintainability.
+2.  **Evaluation of the Solution:**
+    *   **Core Functionality:** The patch correctly extracts the UI elements into well-defined, modular components (`SummaryGrid`, `PrefectureSection`, `VisitCalendar`) and moves the complex state management and calculations into a custom hook (`useStatsData`). It successfully achieves the goal of simplifying the main `StatsPage` component.
+    *   **Safety & Side Effects:**
+        *   All props and states are cleanly passed to the newly created components.
+        *   CSS imports (`../stats.module.css`) correctly resolve from the new `components` subdirectory.
+        *   Global styles for the calendar (`react-calendar/dist/Calendar.css`, `./calendar.css`) were properly left in the main `page.tsx` file, ensuring they still load.
+        *   There is a slight modification where `setTimeout(..., 0)` was introduced inside the `useEffect` hook. While not strictly necessary (as `useEffect` is already asynchronous), it is functionally safe and the cleanup function correctly clears the timeout to prevent memory leaks or state updates on unmounted components.
+        *   The patch includes unrelated updates to `package.json` and `package-lock.json` (moving `babel-plugin-react-compiler` from `devDependencies` to `dependencies`). This is a typical artifact of running `npm install` in some environments. It does not introduce vulnerabilities and will not break the application build, though it is technically out of scope.
+    *   **Completeness:** The refactoring is comprehensive. The entire file was cleaned up and split into logical pieces without breaking the existing architecture.
+3.  **Merge Assessment:**
+    *   The out-of-scope `package.json` changes and the introduction of `setTimeout` inside the hook are **Nitpicks**. They do not harm the application or introduce regressions, but could be cleaned up in a future PR. The core objective of improving code health and readability was accomplished excellently.
+
+### Final Rating: #Correct#
