+++ src/app/stats/page.tsx
@@ -1,66 +1,22 @@
 "use client";
 
-import { useEffect, useMemo, useState } from 'react';
 import Link from 'next/link';
-import Calendar from 'react-calendar';
 import styles from './stats.module.css';
 import 'react-calendar/dist/Calendar.css';
 import './calendar.css';
 import MonthlyVisitsChart from '@/components/charts/MonthlyVisitsChart';
 import RatingDistributionChart from '@/components/charts/RatingDistributionChart';
-import { SaunaVisit } from "@/components/sauna-map/types";
-import {
-  getInitialTheme,
-  flattenVisitHistory,
-  getInitialVisits,
-  calculateStats,
-} from "@/components/sauna-map/utils";
+import { useStatsData } from './hooks/useStatsData';
+import { SummaryGrid } from './components/SummaryGrid';
+import { PrefectureSection } from './components/PrefectureSection';
+import { VisitCalendar } from './components/VisitCalendar';
 
 export default function StatsPage() {
-  const [visits, setVisits] = useState<SaunaVisit[]>([]);
-  const [theme, setTheme] = useState<'dark' | 'light'>('dark');
-  const [date, setDate] = useState<Date | null>(null);
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(() => {
-    setMounted(true);
-    setVisits(getInitialVisits());
-    setTheme(getInitialTheme());
-    setDate(new Date());
-
-    document.documentElement.classList.add("allow-page-scroll");
-    document.body.classList.add("allow-page-scroll");
-    return () => {
-      document.documentElement.classList.remove("allow-page-scroll");
-      document.body.classList.remove("allow-page-scroll");
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!mounted) return;
-    if (theme === 'light') {
-      document.documentElement.classList.add("light-theme");
-    } else {
-      document.documentElement.classList.remove("light-theme");
-    }
-  }, [theme, mounted]);
+  const { visits, theme, date, setDate, mounted, stats, visitDates } = useStatsData();
 
   if (!mounted) {
     return <div className={styles.page} style={{ background: "var(--background)", minHeight: "100vh" }} />;
   }
-
-  const stats = useMemo(() => calculateStats(visits), [visits]);
-
-  const visitDates = useMemo(() => {
-    const dates = new Map<string, number>();
-    flattenVisitHistory(visits)
-      .filter((entry) => entry.status === "visited")
-      .forEach((entry) => {
-        const dateStr = new Date(entry.date).toDateString();
-        dates.set(dateStr, (dates.get(dateStr) ?? 0) + 1);
-      });
-    return dates;
-  }, [visits]);
 
   return (
     <div className={`${styles.page} ${theme === 'light' ? 'light-theme' : ''}`}>
@@ -75,45 +31,9 @@ export default function StatsPage() {
           </Link>
         </header>
 
-        <div className={styles.summaryGrid}>
-          <article className={styles.statCard}>
-            <h3>合計サウナ数</h3>
-            <p>{stats.total}</p>
-          </article>
-          <article className={styles.statCard}>
-            <h3>行った / 行きたい</h3>
-            <p>{stats.visitedCount} / {stats.wishlistCount}</p>
-          </article>
-          <article className={styles.statCard}>
-            <h3>記録エリア数</h3>
-            <p>{stats.uniqueAreas}</p>
-          </article>
-          <article className={styles.statCard}>
-            <h3>平均満足度</h3>
-            <p>{stats.avgRating > 0 ? `${stats.avgRating} / 5` : '-'}</p>
-          </article>
-          <article className={styles.statCard}>
-            <h3>記録期間</h3>
-            <p>{stats.firstDate && stats.lastDate ? `${stats.firstDate} 〜 ${stats.lastDate}` : '-'}</p>
-          </article>
-          <article className={styles.statCard}>
-            <h3>都道府県制覇</h3>
-            <p>{stats.prefectureCount} / 47</p>
-          </article>
-        </div>
+        <SummaryGrid stats={stats} />
 
-        {stats.prefectureCount > 0 && (
-          <section className={styles.prefectureSection}>
-            <h2>都道府県制覇</h2>
-            <div className={styles.badgeList}>
-              {stats.prefectures.map((pref) => (
-                <span key={pref} className={styles.prefectureBadge}>
-                  {pref}
-                </span>
-              ))}
-            </div>
-          </section>
-        )}
+        <PrefectureSection prefectures={stats.prefectures} count={stats.prefectureCount} />
 
         <div className={styles.chartsWrap}>
           <div className={styles.chartGrid}>
@@ -129,37 +49,12 @@ export default function StatsPage() {
           </div>
         </div>
 
-        <div className={styles.chartsWrap}>
-           <section className={styles.chartCard}>
-              <h2>訪問カレンダー</h2>
-              <div className="calendarContainer">
-                <Calendar
-                  onChange={(value) => setDate(value instanceof Date ? value : null)}
-                  value={date}
-                  calendarType="gregory"
-                  className={theme === 'light' ? 'light-theme' : 'dark-theme'}
-                  tileContent={({ date, view }) => {
-                    if (view === 'month') {
-                      const dateStr = date.toDateString();
-                      if (visitDates.has(dateStr)) {
-                        return <div className="calendar-dot"></div>;
-                      }
-                    }
-                    return null;
-                  }}
-                  tileClassName={({ date, view }) => {
-                    if (view === 'month') {
-                      const dateStr = date.toDateString();
-                      if (visitDates.has(dateStr)) {
-                        return "react-calendar__tile--has-visit";
-                      }
-                    }
-                    return null;
-                  }}
-                />
-              </div>
-            </section>
-        </div>
+        <VisitCalendar
+          theme={theme}
+          date={date}
+          setDate={setDate}
+          visitDates={visitDates}
+        />
       </main>
     </div>
   );
